split functional tests of simulators
